Extract request-actor context from browser sessions
