+++ backend/druks/accounts/sessions.py
@@ -1,5 +1,4 @@
 import secrets
-from contextvars import ContextVar
 
 from druks.redis import get_client
 
@@ -7,10 +6,6 @@ from .constants import SESSION_PREFIX
 
 SESSION_COOKIE = "druks_session"
 SESSION_TTL_SECONDS = 30 * 24 * 3600
-
-# The request's signed-in account, stamped by the session gate — Workflow.start
-# reads it so a browser-origin run attributes itself without route ceremony.
-current_account_id: ContextVar[str | None] = ContextVar("current_account_id", default=None)
 
 
 async def mint_session(account_id: str) -> str:
